+++ Owncloud iOs Client/DataBase/Queries/ManageUsersDB.h
@@ -42,7 +42,7 @@
  * Method that return the user object of the idUser
  * @idUser -> id User.
  */
-+ (UserDto *) getUserByIdUser:(int) idUser;
++ (UserDto *) getUserByIdUser:(NSInteger) idUser;
 
 /*
  * Method that return if the user exist or not
